# openapi.yaml
...
openapi: 3.0.3
info:
  title: HyperSDK API
  description: |
    VM migration and export platform providing unified VM export across 9 cloud/virtualization platforms with carbon-aware scheduling.

    Supports: vSphere, Nutanix AHV, AWS, Azure, GCP, Hyper-V, OCI, OpenStack, Alibaba Cloud, Proxmox

    Features:
    - Multi-cloud VM export
    - Carbon-aware scheduling (30-50% carbon reduction)
    - Real-time job monitoring
    - Scheduled backups
  version: 2.0.0
  contact:
    name: HyperSDK
    url: https://github.com/zyvorai/transiva
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.transiva.example.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /jobs/submit:
    post:
      summary: Submit job(s)
      description: Submit one or more VM export jobs
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/JobDefinition'
                - type: array
                  items:
                    $ref: '#/components/schemas/JobDefinition'
                - $ref: '#/components/schemas/BatchJobDefinition'
          application/x-yaml:
            schema:
              oneOf:
                - $ref: '#/components/schemas/JobDefinition'
                - type: array
                  items:
                    $ref: '#/components/schemas/JobDefinition'
      responses:
        '200':
          description: Job(s) submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/query:
    get:
      summary: Query jobs (GET)
      description: Query jobs with optional filters
      tags: [Jobs]
      parameters:
        - name: all
          in: query
          description: Return all jobs
          schema:
            type: boolean
      responses:
        '200':
          description: Jobs query result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
    post:
      summary: Query jobs (POST)
      description: Query jobs with detailed filters
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Jobs query result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
...
  /jobs/cancel:
    post:
      summary: Cancel jobs
      description: Cancel one or more running jobs
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Cancel result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResponse'
...
  /vms/list:
    get:
      summary: List VMs (query params)
      description: List VMs from a provider using query parameters. Use `provider=nutanix` or `provider=vsphere`.
      tags: [VMs]
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [vsphere, nutanix]
        - name: server
          in: query
          schema:
            type: string
        - name: username
          in: query
          schema:
            type: string
        - name: password
          in: query
          schema:
            type: string
        - name: cluster
          in: query
          schema:
            type: string
          description: Nutanix cluster filter
        - name: detailed
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of VMs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: List VMs
      description: List VMs from a provider
      tags: [VMs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VCenterConfig'
      responses:
        '200':
          description: List of VMs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /vms/info:
    get:
      summary: Get VM info (query params)
      description: Get VM details. Use `provider=nutanix&vm={name|uuid}` or vSphere path-based lookup via POST.
      tags: [VMs]
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [vsphere, nutanix]
        - name: vm
          in: query
          schema:
            type: string
        - name: server
          in: query
          schema:
            type: string
        - name: username
          in: query
          schema:
            type: string
        - name: password
          in: query
          schema:
            type: string
      responses:
        '200':
          description: VM information
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Get VM info
      description: Get detailed information about a specific VM
      tags: [VMs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vcenter:
                  $ref: '#/components/schemas/VCenterConfig'
                vm_path:
                  type: string
      responses:
        '200':
          description: VM information
          content:
            application/json:
              schema:
                type: object
  /vms/export:
    post:
      summary: Export VM (Nutanix)
      description: Synchronous Nutanix AHV NFS pickup export. Requires `provider=nutanix`.
      tags: [VMs]
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [nutanix]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vm, output_path]
              properties:
                vm:
                  type: string
                output_path:
                  type: string
                format:
                  type: string
                  enum: [qcow2, raw]
                mounts:
                  type: object
                  additionalProperties:
                    type: string
                  description: Storage container ID/name to local NFS mount path
                resolve_containers:
                  type: boolean
                enable_pipeline:
                  type: boolean
                dry_run:
                  type: boolean
                server:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Export result
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  result:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
...
  /providers/capabilities:
    get:
      summary: Provider export capabilities
      tags: [Providers]
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [vsphere, nutanix]
      responses:
        '200':
          description: Export capabilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  capabilities:
                    type: object
                  registered:
                    type: boolean
                  timestamp:
                    type: string
                    format: date-time
  /vms/shutdown:
    post:
      summary: Shutdown VM
      description: Gracefully shutdown a VM
      tags: [VMs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vcenter:
                  $ref: '#/components/schemas/VCenterConfig'
                vm_path:
                  type: string
      responses:
        '200':
          description: Shutdown initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
  /webhooks/test:
    post:
      summary: Test webhook
      description: Send a test event to a webhook URL
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Webhook test successful
...
  /libvirt/domain:
    get:
      summary: Get domain
      description: Get details of a specific domain
      tags: [Libvirt]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Domain details
  /libvirt/domain/start:
    post:
      summary: Start domain
      description: Start a libvirt domain
      tags: [Libvirt]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Domain started
  /libvirt/domain/shutdown:
    post:
      summary: Shutdown domain
      description: Gracefully shutdown a domain
      tags: [Libvirt]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Domain shutdown initiated
  /libvirt/snapshots:
    get:
      summary: List snapshots
      description: List all snapshots for a domain
      tags: [Libvirt]
      parameters:
        - name: domain
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of snapshots
  /libvirt/snapshot/create:
    post:
      summary: Create snapshot
      description: Create a new snapshot of a domain
      tags: [Libvirt]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: string
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Snapshot created
  /convert/vm:
    post:
      summary: Convert VM
      description: Convert VM using h2kvm
      tags: [H2KVM]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_path:
                  type: string
                output_path:
                  type: string
      responses:
        '200':
          description: Conversion started
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversion_id:
                    type: string
  /convert/status:
    get:
      summary: Get conversion status
      description: Get status of a VM conversion
      tags: [H2KVM]
      parameters:
        - name: conversion_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversion status
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token obtained from /api/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    JobStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
...
    BatchJobDefinition:
      type: object
      required:
        - jobs
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobDefinition'
    QueryRequest:
      type: object
      properties:
        job_ids:
          type: array
          items:
            type: string
        status:
          type: array
          items:
            $ref: '#/components/schemas/JobStatus'
        all:
          type: boolean
          description: Return all jobs
        limit:
          type: integer
          description: Limit number of results
    QueryResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
    SubmitResponse:
      type: object
      properties:
        job_ids:
          type: array
          items:
            type: string
        accepted:
          type: integer
        rejected:
          type: integer
        errors:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    CancelRequest:
      type: object
      required:
        - job_ids
      properties:
        job_ids:
          type: array
          items:
            type: string
    CancelResponse:
      type: object
      properties:
        cancelled:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: string
        errors:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        timestamp:
          type: string
          format: date-time

    # Carbon-Aware Scheduling Schemas
...